:sparkles: [FEAT] - Implementando delete image.

+++ back/src/modules/cars/useCases/Car/removeCarImage/RemoveCarImageUseCase.ts
@@ -2,6 +2,7 @@ import { inject, injectable } from 'tsyringe'
 import { ICarsRepository } from '../../../repositories/Cars/ICarsRepository'
 import { ICarsImagesRepository } from '../../../repositories/CarsImages/ICarsImagesRepository'
 import { IStorageProvider } from '../../../../../shared/container/providers/StorageProvider/IStorageProvider'
+import { AppError } from '../../../../../shared/errors/AppError'
 
 interface IRequest {
   carId: string
@@ -24,8 +25,16 @@ export class RemoveCarImageUseCase {
   }
 
   async execute({ carId, imageId }: IRequest): Promise<void> {
-    await this.storageProvider.deleteImage()
+    if (!imageId) throw new AppError('_id da imagem não enviado')
+
+    const image = await this.carsImagesRepository.findById(imageId)
+
+    if (!image) throw new AppError('Imagem não encontrada')
+
+    await this.storageProvider.deleteImage(image.imageName)
+
     await this.carsRepository.removeImage(carId, imageId)
+
     await this.carsImagesRepository.delete(imageId)
   }
 }
